requirements.txt
python
matplotlib
ipython
jupyter
pandas
numpy
transformers
pytorch
torchvision
torchaudio
cpuonly
scipy
